test_requirements.txt
-r requirements.txt
black==21.6b0
ipdb==0.13.9
pytest==6.2.4
pytest-cov==2.12.1
pytest-factoryboy==2.1.0
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytest-mock==3.6.1
